+++ src/main/java/org/ataraxii/wishlist/service/ItemService.java
@@ -2,6 +2,7 @@ package org.ataraxii.wishlist.service;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.ataraxii.wishlist.database.entity.Folder;
 import org.ataraxii.wishlist.database.entity.Item;
 import org.ataraxii.wishlist.database.entity.ItemFolder;
@@ -11,7 +12,6 @@ import org.ataraxii.wishlist.database.repository.ItemFolderRepository;
 import org.ataraxii.wishlist.database.repository.ItemRepository;
 import org.ataraxii.wishlist.dto.item.ItemDto;
 import org.ataraxii.wishlist.dto.item.ItemResponseDto;
-import org.ataraxii.wishlist.exception.ConflictException;
 import org.ataraxii.wishlist.exception.NotFoundException;
 import org.ataraxii.wishlist.mapper.ItemMapper;
 import org.springframework.stereotype.Service;
@@ -20,6 +20,7 @@ import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class ItemService {
@@ -31,6 +32,7 @@ public class ItemService {
 
     @Transactional
     public ItemResponseDto createItem(ItemDto dto, User currentUser) {
+        log.info("Создание предмета {} пользователем {}", dto.getName(), currentUser.getUsername());
         Item item = Item.builder()
                 .name(dto.getName())
                 .url(dto.getUrl())
@@ -41,9 +43,14 @@ public class ItemService {
 
         if (dto.getFolderId() != null) {
             Folder folder = folderRepository.findByIdAndUser(dto.getFolderId(), currentUser)
-                    .orElseThrow(() -> new NotFoundException("Папка не найдена"));
+                    .orElse(null);
 
+            if (folder == null) {
+                log.warn("Ошибка создания предмета: папка с id={} не найдена у пользователя {}", dto.getFolderId(), currentUser.getUsername());
+                throw new NotFoundException("Папка не найдена");
+            }
             itemRepository.save(item);
+            log.info("Предмет {} успешно создан в папке {} пользователем {}", item.getName(), folder.getName(), currentUser.getUsername());
 
             ItemFolder itemFolder = ItemFolder.builder()
                     .item(item)
@@ -53,12 +60,14 @@ public class ItemService {
             folderId = folder.getId();
         } else {
             itemRepository.save(item);
+            log.info("Предмет {} успешно создан пользователем {}", item.getName(), currentUser.getUsername());
         }
         return itemMapper.toDto(item, folderId);
     }
 
     public List<ItemResponseDto> findAllItems(User currentUser) {
         List<Item> items = itemRepository.findAllByUser(currentUser);
+        log.info("Найдено {} предметов у пользователя {}", items.size(), currentUser.getUsername());
         return items.stream()
                 .map(item -> {
                     UUID folderId = null;
@@ -73,16 +82,29 @@ public class ItemService {
 
     public ItemResponseDto findItemById(User currentUser, UUID id) {
         Item item = itemRepository.findByIdAndUser(id, currentUser)
-                .orElseThrow(() -> new NotFoundException("Предмет с таким id не найден"));
+                .orElse(null);
+        if (item == null) {
+            log.warn("Ошибка поиска предмета: предмет с id={} не найдена у пользователя {}", id, currentUser.getUsername());
+            throw new NotFoundException("Предмет с таким id не найден");
+        }
+        log.info("Предмет {} найден у пользователя {}", item.getName(), currentUser.getUsername());
         return itemMapper.toDto(item);
     }
 
     @Transactional
     public ItemResponseDto updateItem(User currentUser, UUID id, ItemDto updatedItem) {
         Item item = itemRepository.findByIdAndUser(id, currentUser)
-                .orElseThrow(() -> new NotFoundException("Предмет с таким id не найден"));
+                .orElse(null);
+
+        if (item == null) {
+            log.warn("Ошибка обновления предмета: предмет с id={} не найден у пользователя {}", id, currentUser.getUsername());
+            throw new NotFoundException("Предмет с таким id не найден");
+        }
+        log.info("Обновление папки {} пользователем {}", item.getName(), currentUser.getUsername());
+
         item.setName(updatedItem.getName());
         item.setUrl(updatedItem.getUrl());
+        log.info("Предмет {} успешно изменен", item.getName());
 
         itemRepository.save(item);
         return itemMapper.toDto(item);
@@ -91,8 +113,13 @@ public class ItemService {
     @Transactional
     public void deleteItem(User currentUser, UUID id) {
         Item item = itemRepository.findByIdAndUser(id, currentUser)
-                .orElseThrow(() -> new NotFoundException("Предмет с таким id не найден"));
+                .orElse(null);
 
+        if (item == null) {
+            log.warn("Ошибка удаления предмета: предмет с id={} не найден у пользователя {}", id, currentUser.getUsername());
+            throw new NotFoundException("Предмет с таким id не найден");
+        }
         itemRepository.delete(item);
+        log.info("Предмет {} удален пользователем {}", item.getName(), currentUser.getUsername());
     }
 }
